+++ prog/dftb+/lib_timedep/linresp.F90
@@ -153,7 +153,7 @@ contains
 
 
   !> Initialize an internal data type for linear response excitations
-  subroutine LinResp_init(this, ini, nAtom, nEl, orb)
+  subroutine LinResp_init(this, ini, nAtom, nEl, orb, tCasidaForces)
 
     !> data structure for linear response
     type(linresp), intent(out) :: this
@@ -169,6 +169,9 @@ contains
 
     !> data on atomic orbitals
     type(TOrbitals), intent(in) :: orb
+
+    !> Are excited state force contributions required
+    logical, intent(in) :: tCasidaForces
 
 #:if WITH_ARPACK
     this%nExc = ini%nExc
@@ -248,7 +251,7 @@ contains
   !> Wrapper to call the actual linear response routine for excitation energies
   subroutine LinResp_calcExcitations(this, tSpin, denseDesc, eigVec, eigVal, SSqrReal, filling,&
       & coords0, sccCalc, dqAt, species0, iNeighbour, img2CentCell, orb, tWriteTagged, fdTagged,&
-      & excEnergy)
+      & excEnergy, allExcEnergies)
 
     !> data structure with additional linear response values
     type(linresp), intent(inout) :: this
@@ -301,6 +304,9 @@ contains
     !> excitation energy (only when nStat /=0, othewise set numerically 0)
     real(dp), intent(out) :: excEnergy
 
+    !> energes of all solved states
+    real(dp), intent(inout), allocatable :: allExcEnergies(:)
+
 #:if WITH_ARPACK
     @:ASSERT(this%tInit)
     @:ASSERT(size(orb%nOrbAtom) == this%nAtom)
@@ -310,7 +316,7 @@ contains
         & fdTagged, this%fdMulliken, this%fdCoeffs, this%tGrndState, this%fdXplusY, this%fdTrans,&
         & this%fdSPTrans, this%fdTradip, this%tArnoldi, this%fdArnoldi,  this%fdArnoldiDiagnosis,&
         & this%fdExc, this%tEnergyWindow, this%energyWindow, this%tOscillatorWindow,&
-        & this%oscillatorWindow, excEnergy)
+        & this%oscillatorWindow, excEnergy, allExcEnergies)
 
 #:else
     call error('Internal error: Illegal routine call to LinResp_calcExcitations')
@@ -322,8 +328,8 @@ contains
   !> Wrapper to call linear response calculations of excitations and forces in excited states
   subroutine LinResp_addGradients(tSpin, this, iAtomStart, eigVec, eigVal, SSqrReal, filling,&
       & coords0, sccCalc, dqAt, species0, iNeighbour, img2CentCell, orb, skHamCont, skOverCont,&
-      & tWriteTagged, fdTagged, excEnergy, excgradient, derivator, rhoSqr, occNatural,&
-      & naturalOrbs)
+      & tWriteTagged, fdTagged, excEnergy, allExcEnergies, excgradient, derivator, rhoSqr,&
+      & occNatural, naturalOrbs)
 
     !> is this a spin-polarized calculation
     logical, intent(in) :: tSpin
@@ -388,6 +394,9 @@ contains
     !> energy of particular excited state
     real(dp), intent(out) :: excenergy
 
+    !> energes of all solved states
+    real(dp), intent(inout), allocatable :: allExcEnergies(:)
+
     !> contribution to forces from derivative of excited state energy
     real(dp), intent(out) :: excgradient(:,:)
 
@@ -419,8 +428,8 @@ contains
           & this%fdMulliken, this%fdCoeffs, this%tGrndState, this%fdXplusY, this%fdTrans,&
           & this%fdSPTrans, this%fdTradip, this%tArnoldi, this%fdArnoldi, this%fdArnoldiDiagnosis,&
           & this%fdExc, this%tEnergyWindow, this%energyWindow, this%tOscillatorWindow,&
-          & this%oscillatorWindow, excEnergy, shiftPerAtom, skHamCont, skOverCont, excgradient,&
-          & derivator, rhoSqr, occNatural, naturalOrbs)
+          & this%oscillatorWindow, excEnergy, allExcEnergies, shiftPerAtom, skHamCont, skOverCont,&
+          & excgradient, derivator, rhoSqr, occNatural, naturalOrbs)
     else
       call LinRespGrad_old(tSpin, this%nAtom, iAtomStart, eigVec, eigVal, sccCalc, dqAt, coords0,&
           & this%nExc, this%nStat, this%symmetry, SSqrReal, filling, species0, this%HubbardU,&
@@ -428,8 +437,8 @@ contains
           & this%fdMulliken, this%fdCoeffs, this%tGrndState, this%fdXplusY, this%fdTrans,&
           & this%fdSPTrans, this%fdTradip, this%tArnoldi, this%fdArnoldi, this%fdArnoldiDiagnosis,&
           & this%fdExc, this%tEnergyWindow, this%energyWindow, this%tOscillatorWindow,&
-          & this%oscillatorWindow, excEnergy, shiftPerAtom, skHamCont, skOverCont, excgradient,&
-          & derivator, rhoSqr)
+          & this%oscillatorWindow, excEnergy, allExcEnergies, shiftPerAtom, skHamCont, skOverCont,&
+          & excgradient, derivator, rhoSqr)
     end if
 
 #:else