+++ prog/dftb+/prg_dftb/inputdata.F90
@@ -86,6 +86,8 @@ module inputdata_module
     !> Read starting charges from disc
     logical :: tReadChrg = .false.
 
+    logical :: tSkipChrgChecksum = .false.
+
     !> Disc charges are stored as ascii or binary files
     logical :: tReadChrgAscii = .true.
 
@@ -158,7 +160,10 @@ module inputdata_module
     logical :: tProjEigenvecs = .false.
 
     !> Evaluate forces
-    logical :: tForces     = .false.
+    logical :: tForces = .false.
+
+    !> Evaluate force contributions from the excited state if required and (tForces)
+    logical :: tCasidaForces = .false.
 
     !> force evaluation method
     integer :: forceType